moved button to event sidebar, created a location state, and  created a handle change for selecting the location
